Тесты, проверяющие загрузку файла, вынесены в отдельный файл
